feat: implementing MongoDB Repos and Connection

+++ src/controller/message.ts
@@ -1,19 +1,38 @@
 import { Message } from "../model/messages";
-
 import { MessageRepository } from "../repository/messageRepository";
+import { NotFoundError } from "../Error/notFoundError";
+import { BadRequestError } from "../Error/badRequestError";
 
 export class MessageController {
   constructor(public messageRepository: MessageRepository) {}
 
-  createMessage(message: Message) {
-    return this.messageRepository.createMessage(message);
+  async createMessage(message: Message) {
+    if (!message) {
+      throw new BadRequestError("Please provide a message");
+    }
+    const NewMessage = this.messageRepository.createMessage(message);
+    if (!NewMessage) {
+      throw new BadRequestError("Error while creating message");
+    }
+    return NewMessage;
   }
 
-  getMessages() {
-    return this.messageRepository.getMessages();
+  async getMessagesByChatId(chatId: string) {
+    if (!chatId) {
+      throw new BadRequestError("Please provide a chatId");
+    }
+    const messages = this.messageRepository.getMessagesByChatId(chatId);
+    if (!messages) {
+      throw new NotFoundError("No messages found");
+    }
+    return messages;
   }
 
-  deleteMessage(id: string) {
+  async deleteMessage(id: string) {
+    const currentMessage = await this.messageRepository.getMessageById(id);
+    if (!currentMessage) {
+      throw new NotFoundError("Message not found");
+    }
     return this.messageRepository.deleteMessage(id);
   }
 }
